+++ src/token.js
@@ -5,12 +5,19 @@ const { Util } = require('./jsotp/util');
 
 const TOKEN_DELIMITER = '::';
 const TOTP_INTERVAL = 15;
-const VALID_BARCODE_REGEX = /^[0-9]{12,18}[A-Za-z]?$/;
+const RAW_BARCODE_REGEX = /^[0-9]{12,18}[A-Za-z]?$/;
 
-const SegmentType = {
+const RenderType = {
     BARCODE: 'BARCODE',
     ROTATING_SYMBOLOGY: 'ROTATING_SYMBOLOGY',
     UNKNOWN: 'UNKNOWN'
+};
+
+const DisplayType = {
+    INVALID: 'INVALID',
+    STATIC_QR: 'STATIC_QR',
+    STATIC_PDF: 'STATIC_PDF',
+    ROTATING: 'ROTATING'
 };
 
 // We use WeakMaps to mimic private access modifier.
@@ -18,24 +25,26 @@ const _barcode = new WeakMap();
 const _rawToken = new WeakMap();
 const _customerKey = new WeakMap();
 const _eventKey = new WeakMap();
-const _segmentType = new WeakMap();
+const _renderType = new WeakMap();
+const _displayType = new WeakMap();
 
 /**
  * Class to sign token's supplied from Presence Delivery Service.
  */
 class TokenSigner {
     constructor(encodedToken) {
-        let json = { b: 'errorbarcode', segmentType: SegmentType.UNKNOWN };
+        let json = { b: 'errorbarcode', rt: RenderType.UNKNOWN };
 
         try {
             const decodedToken = global.atob(encodedToken);
             json = JSON.parse(decodedToken);
+            if (json.rt) {
+                json.rt = json.rt.toUpperCase();
+            }
         } catch (e) {
-            // Check if we were actually provided a raw barcode, before
-            // continuing with assumption we have an invalid token.
-            const barcode = encodedToken && encodedToken.match(VALID_BARCODE_REGEX);
-            if (barcode) {
-                json = { b: barcode[0], segmentType: SegmentType.BARCODE };
+            const rawBarcode = encodedToken && encodedToken.match(RAW_BARCODE_REGEX);
+            if (rawBarcode) {
+                json = { b: rawBarcode[0], rt: RenderType.BARCODE };
             }
         }
 
@@ -45,14 +54,21 @@ class TokenSigner {
         //    actually have these values?
         //  - What should we do if we're missing something?
         if (json.t && json.ck) {
-            json.segmentType = SegmentType.ROTATING_SYMBOLOGY;
             _rawToken.set(this, json.t);
             _customerKey.set(this, json.ck);
             _eventKey.set(this, json.ek);
         }
 
         _barcode.set(this, json.b);
-        _segmentType.set(this, json.segmentType || SegmentType.BARCODE);
+        _renderType.set(this, json.rt);
+
+        let displayType = DisplayType.INVALID;
+        if ((json.rt === RenderType.BARCODE)) {
+            displayType = DisplayType.STATIC_QR;
+        } else if (json.rt === RenderType.ROTATING_SYMBOLOGY || !json.rt) {
+            displayType = (json.t) ? DisplayType.ROTATING : DisplayType.STATIC_PDF;
+        }
+        _displayType.set(this, displayType);
     }
 
     get barcode() {
@@ -71,16 +87,21 @@ class TokenSigner {
         return _eventKey.get(this);
     }
 
-    get segmentType() {
-        return _segmentType.get(this);
+    get renderType() {
+        return _renderType.get(this);
+    }
+
+    get displayType() {
+        return _displayType.get(this);
     }
 
     /**
-     * Generates a signed token if this instance's `segmentType` is
-     * ROTATING_SYMBOLOGY, otherwise return the barcode string.
+     * Generates a signed token if this instance's `renderType` is
+     * ROTATING_SYMBOLOGY and we have an encoded token, otherwise return the
+     * barcode string.
      */
     generateSignedToken(time, withCounterPadding = true) {
-        if (this.segmentType === SegmentType.ROTATING_SYMBOLOGY) {
+        if (this.displayType === DisplayType.ROTATING) {
             const keys = [this.eventKey, this.customerKey];
             const tokenComponents = keys.reduce((acc, key) => {
                 if (key) {
@@ -103,4 +124,4 @@ class TokenSigner {
     }
 }
 
-module.exports = { TokenSigner, SegmentType };
+module.exports = { TokenSigner, DisplayType };