+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.1] - 2019-03-22
+### Added
+- Optional parameter to `setToken` allowing customization of token parsing errors.
+- `showError` method that allows displaying an error to the end user when there are issue retrieving
+  a token from the Secure Renderer API.
+
+### Deprecated
+- `SecureEntryView`'s `setErrorText(errorText)` method and `errorText` constructor option that allowed
+  customization of token parsing errors. Use the less ambiguous `setToken(token, parseErrorText)` method
+  instead.
+
 ## [1.0.0] - 2019-02-25
 ### Added
 - AMD module support.